# torque-api-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Torque API Reference",
    "description": "This page contains information about the Torque APIs and how to use it.<br><br>To be able to test the API methods, you will need an access token to be set in the Authorize section.If you got to this page from your Torque account, we already made it available for you, but you can switch to a different token as needed.<br>To get a different token, use the 'Get token' function available under the 'Access Tokens' section. After running the method, copy the access_token you get in the response and set it in the Authorize fieldas 'Bearer access_token'. For example: Bearer fqSWuw72BbUVFn8AbokF77GJ0r5KEn9MiZjLXF8kBwI.<br><br>The API can be accessed with your account endpoint as will be mentioned in the descriptions and examples or using https://qtorque.io as the prefix instead."
  },
  "paths": {
...
    "/api/spaces/{space_name}/blueprints": {
      "get": {
        "tags": [
          "Blueprints"
        ],
        "summary": "Get published blueprints",
        "description": "Returns a list of published blueprints in the space, including sample blueprints.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the desired blueprints",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "Blueprint type: \"sandbox\" or \"production\"<br />\r\n            Optional: default is \"sandbox\"",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commit",
            "in": "query",
            "description": "Blueprint commit: any commit name in a related repository<br />\r\n            Optional: default is null",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "description": "Blueprint branch: any branch name in a related repository<br />\r\n            Optional: default is null",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlueprintForGetAllResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Space not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_name}/blueprints/{blueprint_name}/policies": {
      "get": {
        "tags": [
          "Blueprints"
        ],
        "summary": "Get blueprint policy",
        "description": "Get blueprint policy.\r\nRelevant only for blueprints in repository for sandbox environment.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the desired blueprint",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "blueprint_name",
            "in": "path",
            "description": "Name of the blueprint",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintPolicyResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space or blueprint not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Blueprint belongs to production repository",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Blueprints"
        ],
        "summary": "Update blueprint policy",
        "description": "Update policies of a blueprint. Relevant only for blueprints in a sandbox repository.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space where the blueprint resides",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "blueprint_name",
            "in": "path",
            "description": "Blueprint name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update blueprint policy request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBlueprintPolicyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "422": {
            "description": "Required parameters are missing, max duration too short or blueprint belongs to production repository",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space or blueprint not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Blueprints"
        ],
        "summary": "Reset blueprint policies",
        "description": "Reset blueprint policies to the default behavior. Max duration will be cleared.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space where the blueprint resides",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "blueprint_name",
            "in": "path",
            "description": "Blueprint name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Space or blueprint not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Blueprint belongs to production repository",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/spaces/{space_name}/validations/blueprints": {
      "post": {
        "tags": [
          "Blueprints"
        ],
        "summary": "Validate blueprint",
        "description": "Runs validation check on a specific blueprint in the space.",
        "parameters": [
          {
            "name": "space_name",
            "in": "path",
            "description": "Name of the space containing the blueprint",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlueprintValidationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlueprintValidationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Space or blueprint repository not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Blueprint name, blueprint type, or blueprint repository branch is empty",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BlueprintApplicationForGetAllResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Application name",
            "nullable": true
          },
          "version": {
            "type": "string",
            "description": "Application version",
            "nullable": true
          },
          "icon_url": {
            "type": "string",
            "description": "URL of the application icon",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Blueprint application information",
        "x-internal": true
      },
      "BlueprintCostResponse": {
        "type": "object",
        "properties": {
          "average_hourly_cost": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false,
        "x-internal": true
      },
      "BlueprintForGetAllResponse": {
        "type": "object",
        "properties": {
          "blueprint_name": {
            "type": "string",
            "description": "Blueprint name",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Blueprint description",
            "nullable": true
          },
          "url": {
            "type": "string",
            "description": "Blueprint repository URL",
            "nullable": true
          },
          "inputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintInputResponse"
            },
            "nullable": true
          },
          "enabled": {
            "type": "boolean",
            "description": "Indicates if the blueprint is published or not"
          },
          "last_modified": {
            "type": "string",
            "description": "Last modification date and time of the blueprint",
            "format": "date-time"
          },
          "modified_by": {
            "type": "string",
            "description": "Name of the last user who modified the blueprint",
            "nullable": true
          },
          "applications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintApplicationForGetAllResponse"
            },
            "nullable": true
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintServiceForGetAllResponse"
            },
            "nullable": true
          },
          "clouds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CloudResponse"
            },
            "nullable": true
          },
          "is_sample": {
            "type": "boolean",
            "description": "Indicates if the blueprint is a sample blueprint"
          },
          "artifacts": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Artifact files of the blueprint applications\r\n<remarks>Maps application name to its relative artifacts path in the space</remarks>",
            "nullable": true
          },
          "environment_type": {
            "$ref": "#/components/schemas/EnvironmentType"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            },
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintTagResponse"
            },
            "nullable": true
          },
          "policies": {
            "$ref": "#/components/schemas/BlueprintPolicyResponse"
          },
          "cost": {
            "$ref": "#/components/schemas/BlueprintCostResponse"
          },
          "links": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Hyperlink"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Blueprint information",
        "x-internal": true
      },
      "BlueprintInputResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Input name",
            "nullable": true
          },
          "default_value": {
            "type": "string",
            "description": "When the sandbox is created, Torque automatically populates the input with this value.\r\n<remarks>\r\nUser can choose to edit the value or leave it as-is\r\n</remarks>",
            "nullable": true
          },
          "display_style": {
            "type": "string",
            "description": "Indicates how to display the input value in UI\r\n<remarks>\r\n- To display the input value in plain text in the UI, do not assign a value.<br />\r\n- To hide the input value behind bullets, enter the value \"masked\".\r\n</remarks>",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Input description",
            "nullable": true
          },
          "optional": {
            "type": "boolean",
            "description": "Indicates if the input is optional or not\r\n<remarks>\r\n- When optional is set to true, the user can leave the parameter empty.<br />\r\n- When optional is set to false, empty value(s) will result in validation error(s).\r\n</remarks>"
          },
          "possible_values": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Possible values for the input\r\n<remarks>\r\n- Possible_values are optional\r\n- The default value is one of the possible values\r\n- Possible values must be unique\r\n</remarks>",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Blueprint input information",
        "x-internal": true
      },
      "BlueprintPolicyResponse": {
        "type": "object",
        "properties": {
          "max_duration": {
            "type": "string",
            "description": "Max duration of sandbox in ISO 8601 format: P{days}DT{hours}H{minutes}M{seconds}S] (for example, 'P0DT2H3M4S')",
            "format": "date-span",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Blueprint policy response",
        "x-internal": true
      },
      "BlueprintServiceForGetAllResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Service name\r\n<remarks>A service folder and YAML with this name should reside in the \"/services\" folder in the blueprint YAML’s repository</remarks>",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Service type (e.g. \"Terraform\")",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Blueprint service information",
        "x-internal": true
      },
      "BlueprintSourceRequest": {
        "type": "object",
        "properties": {
          "branch": {
            "type": "string",
            "description": "Blueprint repository branch",
            "nullable": true
          },
          "commit": {
            "type": "string",
            "description": "Blueprint commit version",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request containing blueprint repository information",
        "x-internal": true
      },
      "BlueprintSourceResponse": {
        "type": "object",
        "properties": {
          "branch": {
            "type": "string",
            "description": "Blueprint repository branch",
            "nullable": true
          },
          "commit": {
            "type": "string",
            "description": "Blueprint commit version",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Blueprint repository information",
        "x-internal": true
      },
      "BlueprintTagResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "default_value": {
            "type": "string",
            "nullable": true
          },
          "possible_values": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "x-internal": true
      },
      "BlueprintValidationRequest": {
        "type": "object",
        "properties": {
          "blueprint_name": {
            "type": "string",
            "description": "Blueprint name",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Blueprint type: \"sandbox\" or \"production\"",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/BlueprintSourceRequest"
          }
        },
        "additionalProperties": false,
        "description": "Request to validate blueprint",
        "x-internal": true
      },
      "BlueprintValidationResponse": {
        "type": "object",
        "properties": {
          "blueprint_name": {
            "type": "string",
            "description": "Blueprint name",
            "nullable": true
          },
          "description": {
            "type": "string",
            "description": "Blueprint description",
            "nullable": true
          },
          "url": {
            "type": "string",
            "description": "Blueprint repository URL",
            "nullable": true
          },
          "source": {
            "$ref": "#/components/schemas/BlueprintSourceResponse"
          },
          "inputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintInputResponse"
            },
            "nullable": true
          },
          "applications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintApplicationForGetAllResponse"
            },
            "nullable": true
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlueprintServiceForGetAllResponse"
            },
            "nullable": true
          },
          "clouds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CloudResponse"
            },
            "nullable": true
          },
          "artifacts": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "description": "Artifact files of blueprint applications\r\n<remarks>Maps application name to its relative artifacts path in the space</remarks>",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Blueprint validation response",
        "x-internal": true
      },
...
      "CloudComputeServiceResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Cloud compute service name",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Cloud compute service type (EC2, Azure virtual machine, AKS, EKS, etc.)",
            "nullable": true
          },
          "created_date": {
            "type": "string",
            "description": "Cloud compute service creation date and time",
            "format": "date-time"
          },
          "created_by": {
            "type": "string",
            "description": "User name of the cloud compute service owner",
            "nullable": true
          },
          "user_defined": {
            "type": "boolean",
            "description": "Indicates if the cloud compute service created automatically or defined by the user<br /><example>EC2 and Azure VM are not user defined while K8s, EKS and AKS are user-defined.</example>"
          }
        },
        "additionalProperties": false,
        "description": "Cloud compute service information",
        "x-internal": true
      },
      "CloudRegionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Cloud region ID",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Cloud region name",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Cloud region information",
        "x-internal": true
      },
      "CloudResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "description": "Cloud provider name",
            "nullable": true
          },
          "cloud_account_name": {
            "type": "string",
            "description": "Cloud account name",
            "nullable": true
          },
          "compute_service": {
            "$ref": "#/components/schemas/CloudComputeServiceResponse"
          },
          "region": {
            "$ref": "#/components/schemas/CloudRegionResponse"
          }
        },
        "additionalProperties": false,
        "description": "Cloud account or Kubernetes compute service information",
        "x-internal": true
      },
...
      "EnvironmentType": {
        "type": "object",
        "additionalProperties": false,
        "x-internal": true
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message",
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Error name",
            "nullable": true
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Error information",
        "x-internal": true
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Error"
            },
            "description": "Array of errors",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Errors received from an API",
        "x-internal": true
      },
      "Hyperlink": {
        "type": "object",
        "properties": {
          "rel": {
            "type": "string",
            "description": "Link relation type that describes how the current context (source) is related to the target resource",
            "nullable": true
          },
          "href": {
            "type": "string",
            "description": "URI link to the API",
            "nullable": true
          },
          "method": {
            "type": "string",
            "description": "HTTP method (GET, POST, PUT, etc.)",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Links to API",
        "x-internal": true
      },
...
      "UpdateBlueprintPolicyRequest": {
        "type": "object",
        "properties": {
          "max_duration": {
            "type": "string",
            "description": "Max duration of sandbox in ISO 8601 format: P{days}DT{hours}H{minutes}M{seconds}S] (for example, 'P0DT2H3M4S')",
            "format": "date-span",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Update Blueprint policy request",
        "x-internal": true
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "Authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}